Move requestContext into runtime/
The AsyncLocalStorage request-scoped state (runtime token, session id, wiki
key) lived under transport/, but it is a runtime primitive consumed by runtime/
(dispatcher, wikiCapability) and wikis/ (state, activeWiki) as well as the
transport. Those consumers therefore imported "up" into transport — an inverted
layer edge. Move it to runtime/ so every consumer imports down into a lower
layer, and repoint the imports.

Also drop streamableHttp's re-export of withRequestContext: it existed only for
one test, which now imports the symbol from its real home alongside the other
requestContext helpers. Pure move, no behaviour change.
