+++ src/runtime/dispatcher.ts
@@ -4,7 +4,7 @@ import type { Tool } from './tool.js';
 import type { ToolContext } from './context.js';
 import { applySpecialCase } from '../errors/specialCases.js';
 import { errorMessage } from '../errors/isErrnoException.js';
-import { getRuntimeToken, getSessionId, withRequestFields } from '../transport/requestContext.js';
+import { getRuntimeToken, getSessionId, withRequestFields } from './requestContext.js';
 import { isWikiScoped, normalizeWikiArg } from './wikiArg.js';
 import {
 	emitToolCall,
