+++ src/runtime/wikiCapability.ts
@@ -2,7 +2,7 @@ import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 import type { ToolContext } from './context.js';
 import type { ExtensionPack } from '../tools/extensions/types.js';
 import { extensionPacks } from '../tools/extensions/index.js';
-import { getRuntimeToken } from '../transport/requestContext.js';
+import { getRuntimeToken } from './requestContext.js';
 import { hasStaticCredentials } from '../transport/bearerGuard.js';
 
 const CORE_WRITE_TOOL_NAMES: readonly string[] = [
