+++ src/transport/streamableHttp.ts
@@ -24,7 +24,7 @@ import {
 	setSessionsProvider,
 	setProxyStoreStatsProvider,
 } from '../runtime/metrics.js';
-import { withRequestContext } from './requestContext.js';
+import { withRequestContext } from '../runtime/requestContext.js';
 import {
 	createInFlightCounter,
 	markSessionActive,
@@ -32,8 +32,6 @@ import {
 	type SessionRegistry,
 	type InFlightCounter,
 } from './sessionRegistry.js';
-
-export { withRequestContext } from './requestContext.js';
 import { loadConfigFromFile, type WikiConfig } from '../config/loadConfig.js';
 import type { MwnProvider } from '../wikis/mwnProvider.js';
 import type { ActiveWiki } from '../wikis/activeWiki.js';
