+++ tests/runtime/wikiCapability.stepup.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { checkWikiCapability } from '../../src/runtime/wikiCapability.js';
 import { fakeContext } from '../helpers/fakeContext.js';
-import { withRequestFields } from '../../src/transport/requestContext.js';
+import { withRequestFields } from '../../src/runtime/requestContext.js';
 import type { ProxyConfig } from '../../src/auth/authorizationServer/proxyConfig.js';
 
 const PROXY = {
